Move minimal player errors reducer to playlist.live.playerErrors
